GNU/Linux: Improve resolving font filepath
I don't know if this case is possible, but we should have such a
logic just in case.

Windows:

`GetModuleFileNameW()` returns the fully qualified path, so there
is no need to assume such a case.

macOS:

We assume the app bundle structure, so even if
`_NSGetExecutablePath()` returns a relative path, it should be at
least the following hierarchical structure. So there is no need
to assume such a case.

    RaylibIMEInputSampleApp.app/Contents/MacOS/RaylibIMEInputSampleApp